refactor error screen
- move error summary into component
- fix css
- use class for rotation
