+++ src/main/java/digital/slovensko/autogram/core/FailedVisualizationException.java
@@ -5,6 +5,10 @@ import digital.slovensko.autogram.ui.gui.IgnorableException;
 public class FailedVisualizationException extends IgnorableException {
 
     public FailedVisualizationException(Throwable e, SigningJob job, Runnable onContinueCallback) {
-        super("Neočakávaná chyba", "Nastala neznáma chyba", "Pri zobrazovaní dokumentu nastala neočakávaná chyba.", e, job, onContinueCallback);
+        super(
+            "Pri zobrazovaní dokumentu nastala chyba.",
+            "Chcete pokračovať v podpisovaní?", 
+                "Stále je možné dokument podpísať, ale uistite sa, že dôverujete zdroju dokumentu.\nKontaktujte správcu systému a nahláste mu chybu.",
+                e, job, onContinueCallback);
     }
 }
