+++ src/main/resources/digital/slovensko/autogram/ui/gui/error-dialog.fxml
@@ -7,20 +7,17 @@
 <?import javafx.scene.text.TextFlow?>
 
 <VBox xmlns="http://javafx.com/javafx"
-      xmlns:fx="http://javafx.com/fxml"
-      prefWidth="450" minWidth="450"
-      fx:id="mainBox">
+    xmlns:fx="http://javafx.com/fxml"
+    prefWidth="450" minWidth="450"
+    fx:id="mainBox">
 
-    <VBox styleClass="autogram-error-summary">
-        <TextFlow styleClass="autogram-error-summary__title"><Text fx:id="heading">Nastala chyba</Text></TextFlow>
-        <TextFlow styleClass="autogram-error-summary__error"><Text fx:id="subheading">Miesto pre nazov informacie</Text></TextFlow>
-        <TextFlow styleClass="autogram-error-summary__description"><Text fx:id="description">description</Text></TextFlow>
-    </VBox>
+    <fx:include fx:id="errorSummaryComponent" source="error-summary-component.fxml" />
 
     <HBox styleClass="autogram-actions">
-        <Button fx:id="mainButton" styleClass="autogram-button,autogram-button--secondary" text="Pokračovať" onAction="#onMainButtonAction"/>
-        <Button fx:id="showErrorDetailsButton" styleClass="autogram-link" text="Zobraziť detail chyby" visible="false" onAction="#onShowErrorDetailsButtonAction" />
+        <Button fx:id="mainButton" styleClass="autogram-button,autogram-button--secondary"
+            text="Pokračovať" onAction="#onMainButtonAction" />
     </HBox>
 
-    <TextArea fx:id="errorDetails" styleClass="autogram-details__text" editable="false" visible="false" managed="false">Error details</TextArea>
+    <TextArea fx:id="errorDetails" styleClass="autogram-details__text" editable="false"
+        visible="false" managed="false">Error details</TextArea>
 </VBox>
