+++ src/main/resources/digital/slovensko/autogram/ui/gui/ignorable-exception-dialog.fxml
@@ -5,33 +5,13 @@
 
 <?import javafx.scene.text.Text?>
 <?import javafx.scene.text.TextFlow?>
-<?import javafx.scene.shape.Polygon?>
 
 <VBox xmlns="http://javafx.com/javafx"
     xmlns:fx="http://javafx.com/fxml"
     prefWidth="450" minWidth="450"
     fx:id="mainBox">
 
-    <VBox styleClass="autogram-error-summary">
-        <TextFlow styleClass="autogram-error-summary__title"><Text fx:id="heading">Nastala chyba</Text></TextFlow>
-        <TextFlow styleClass="autogram-error-summary__error"><Text fx:id="subheading">Miesto pre nazov informacie</Text></TextFlow>
-        <TextFlow styleClass="autogram-error-summary__description,autogram-body-s"><Text fx:id="description">description</Text></TextFlow>
-
-        <TextFlow styleClass="autogram-body">
-            <Button fx:id="showErrorDetailsButton"
-                styleClass="autogram-link,autogram-error-summary__more-standalone"
-                text="Zobraziť detail chyby"
-                onAction="#onShowErrorDetailsButtonAction">
-                <graphic>
-                    <Polygon points="0.0, 0.0, 12.0, 6.0, 0.0, 10.0">
-                    </Polygon>
-                </graphic>
-            </Button>
-        </TextFlow>
-
-        <TextArea fx:id="errorDetails" styleClass="autogram-details__text" editable="false"
-            visible="false" managed="false">Error details</TextArea>
-    </VBox>
+    <fx:include fx:id="errorSummaryComponent" source="error-summary-component.fxml" />
 
     <HBox styleClass="autogram-actions">
         <Button fx:id="cancelButton" styleClass="autogram-button,autogram-button--warning"
